Rest cart endPoint service

+++ Green-Shop-Rest/src/main/java/com/example/greenshoprest/service/impl/OrderServiceImpl.java
@@ -1,8 +1,11 @@
 package com.example.greenshoprest.service.impl;
 
+import com.example.greenshopcommon.dto.orderDto.CreateOrderRequestDto;
 import com.example.greenshopcommon.dto.orderDto.OrderDto;
 import com.example.greenshopcommon.dto.productDto.ProductDto;
+import com.example.greenshopcommon.entity.Cart;
 import com.example.greenshopcommon.entity.Order;
+import com.example.greenshopcommon.entity.Product;
 import com.example.greenshopcommon.entity.User;
 import com.example.greenshopcommon.mapper.CategoryMapper;
 import com.example.greenshopcommon.mapper.OrderMapper;
@@ -12,12 +15,15 @@ import com.example.greenshopcommon.repository.CartRepository;
 import com.example.greenshopcommon.repository.OrderRepository;
 import com.example.greenshopcommon.repository.ProductRepository;
 import com.example.greenshopcommon.repository.UserRepository;
+import com.example.greenshoprest.service.CartService;
 import com.example.greenshoprest.service.OrderService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -32,10 +38,10 @@ public class OrderServiceImpl implements OrderService {
     private final ProductMapper productMapper;
     private final CategoryMapper categoryMapper;
     private final CartRepository cartRepository;
-//    private final CartService cartService;
+    private final CartService cartService;
 
     @Override
-    public OrderDto getOrderById(int id) {
+    public ResponseEntity<OrderDto> getOrderById(int id) {
         return orderRepository.findById(id)
                 .map(order -> {
                     OrderDto orderDto = orderMapper.mapToDto(order);
@@ -47,52 +53,61 @@ public class OrderServiceImpl implements OrderService {
                                 productDto.setCategoryDto(categoryMapper.mapToDto(product.getCategory()));
                                 orderDto.setProductDto(productDto);
                             });
-                    return orderDto;
+                    return ResponseEntity.ok(orderDto);
                 })
-                .orElse(null);
+                .orElse(ResponseEntity.notFound().build());
     }
 
-
     @Override
-    public List<OrderDto> getOrdersByUserId(User user) {
+    public ResponseEntity<List<OrderDto>> getOrdersByUserId(User user) {
         List<Order> allByUserId = orderRepository.findAllByUserId(user.getId());
-        List<OrderDto> allOrderdtos = new ArrayList<>();
-        allByUserId.forEach(order -> {
+        List<OrderDto> allOrderDtos = new ArrayList<>();
+        for (Order order : allByUserId) {
             OrderDto orderDto = orderMapper.mapToDto(order);
-            orderDto.setUserDto(userMapper.mapToDto(order.getUser()));
+            orderDto.setUserDto(userMapper.mapToDto(user));
             ProductDto productDto = productMapper.mapToDto(order.getProduct());
             productDto.setCategoryDto(categoryMapper.mapToDto(order.getProduct().getCategory()));
             orderDto.setProductDto(productDto);
-            allOrderdtos.add(orderDto);
-        });
-        return allOrderdtos;
+            allOrderDtos.add(orderDto);
+        }
+        if (allOrderDtos.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        } else {
+            return ResponseEntity.ok(allOrderDtos);
+        }
     }
-
-//    @Override
-//    public OrderDto addOrder(CreateOrderRequestDto createOrderRequestDto, User user) {
-//        Optional<User> byId = userRepository.findById(user.getId());
-//        Optional<Product> product = productRepository.findById(createOrderRequestDto.getProductDto().getId());
-//        if (byId.isPresent() && product.isPresent()) {
-//            Order order = orderMapper.map(createOrderRequestDto);
-//            order.setUser(byId.get());
-//            order.setProduct(product.get());
-//            orderRepository.save(order);
-//            Optional<Cart> byProductId = cartRepository.findByProductId(product.get().getId());
-//            byProductId.ifPresent(cart -> {
-//                Product cartProduct = cart.getProduct();
-//                cartProduct.setQuantity(cartProduct.getQuantity() - cart.getQuantity());
-//                productRepository.save(cartProduct);
-//                cartService.deleteById(cart.getId());
-//            });
-//            return orderMapper.mapToDto(order);
-//        }
-//        return null;
-//    }
 
 
     @Override
-    public void deleteOrderById(int id) {
+    public ResponseEntity<OrderDto> addOrder(CreateOrderRequestDto createOrderRequestDto, User user) {
+        Optional<User> byId = userRepository.findById(user.getId());
+        Optional<Product> product = productRepository.findById(createOrderRequestDto.getProductDto().getId());
+        if (byId.isPresent() && product.isPresent()) {
+            Order order = orderMapper.map(createOrderRequestDto);
+            order.setUser(byId.get());
+            order.setProduct(product.get());
+            orderRepository.save(order);
+            Optional<Cart> byProductId = cartRepository.findByProductId(product.get().getId());
+            byProductId.ifPresent(cart -> {
+                Product cartProduct = cart.getProduct();
+                cartProduct.setQuantity(cartProduct.getQuantity() - cart.getQuantity());
+                productRepository.save(cartProduct);
+                cartService.deleteById(cart.getId());
+            });
+            return ResponseEntity.ok(orderMapper.mapToDto(order));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    @Override
+    public ResponseEntity<?> deleteOrderById(int id) {
+        Optional<Order> byId = orderRepository.findById(id);
+        if (byId.isEmpty()){
+            return ResponseEntity.noContent().build();
+        }
         orderRepository.deleteById(id);
+        return ResponseEntity.notFound().build();
     }
 
 }
